Refactor: Organize file structure following industry best practices
- Moved repository and interface to flat structure (consistent with single files)
- Created dto/ folder for multiple DTO files (industry standard)
- Added proper DTOs with validation decorators
- Added index.ts for clean imports
- Added UserResponseDto for API responses

Final structure follows NestJS conventions:
- Single files: module level (entity, service, controller, repository)
- Multiple files: dedicated folder (dto/)

All 38 tests passing
