+++ apps/backend/src/users/dto/index.ts
@@ -1,0 +1,3 @@
+export { CreateUserDto } from './create-user.dto';
+export { UpdateUserDto } from './update-user.dto';
+export { UserResponseDto } from './user-response.dto';
